avoid issues around GLdouble in GLES2/ARM boards.

+++ src/core/StelOpenGL.hpp
@@ -196,7 +196,7 @@ int checkGLErrors(const char *file, int line);
 #  define GL_DOUBLE GL_FLOAT
 # endif
 # ifndef GLdouble
-typedef GLfloat GLdouble;
+typedef double GLdouble;
 # endif
 #endif
